Add a new known dialyzer warning
Dialyzer complains about the yaws_config not being able to match the
atom 'undefined' when it calls yaws_generated functions; this is
because yaws_generated is generated code, as its name implies, so it
handles constants as if they were variables. Ignore this complaint by
adding it to known_dialyzer_warnings.

Also in known_dialyzer_warnings, augment each regular expression to
optionally match a column number.