Regenerate TLS snakeoil cert
The erlang emulator complains about not having enough security when
using the TLS snakeoil (i.e. self-signed) cert shipped. This is because
SHA-1 is used for signing algorithm. Nothing in the TLS standard states
that it shouldn't be allowed, but the erlang emulator doesn't allow it.

The TLS snakeoil cert is regenerated using SHA-256 as signing algorithm.