Prevent CRLF injections in redirects
Redirect mappings specified in configuration are subject to CRLF
injections, where the path to be redirected to on another domain is
copied in the Location header, including unencoded \r and \n
characters. This could lead to unwanted headers being added in the
server reply, for example, a Set-cookie directive.

Example in conf:

<server domain1>
 <redirect>
  / = 301 https://domain2
 </redirect>
</server>

Now, going to

https://domain1/%0D%0ASet-Cookie:crlfinjection=crlfinjection

would generate this reply:

HTTP 1.1 301 Moved Permanently
Location: https://domain2/
Set-Cookie: crlfinjection=crlfinjection
...

Encode \r and \n characters as normal characters to prevent injection.