Remove almost all cases of yaws.hyber.org
For years yaws.hyber.org hosted the Yaws website, and so it ended up
being mentioned in various tests, examples, and documentation. But
Klacke retired about 5 years ago and the yaws.hyber.org site retired
with him, so change almost all references to it to use something else,
usually www.example.org. The only remaining references are in a couple
cygwin and gentoo files, which are no longer maintained, and for
historical reasons, in the www/news file.

Fixes #493